+++ backend/canisters/community/impl/src/lib.rs
@@ -90,10 +90,6 @@ impl RuntimeState {
         self.env.caller() == self.data.group_index_canister_id
     }
 
-    pub fn is_caller_local_group_index(&self) -> bool {
-        self.env.caller() == self.data.local_group_index_canister_id
-    }
-
     pub fn is_caller_proposals_bot(&self) -> bool {
         self.env.caller() == self.data.proposals_bot_user_id.into()
     }
@@ -347,7 +343,6 @@ impl RuntimeState {
                 user_index: self.data.user_index_canister_id,
                 group_index: self.data.group_index_canister_id,
                 local_user_index: self.data.local_user_index_canister_id,
-                local_group_index: self.data.local_group_index_canister_id,
                 notifications: self.data.notifications_canister_id,
                 proposals_bot: self.data.proposals_bot_user_id.into(),
                 escrow: self.data.escrow_canister_id,
@@ -414,7 +409,6 @@ struct Data {
     user_index_canister_id: CanisterId,
     local_user_index_canister_id: CanisterId,
     group_index_canister_id: CanisterId,
-    local_group_index_canister_id: CanisterId,
     notifications_canister_id: CanisterId,
     proposals_bot_user_id: UserId,
     escrow_canister_id: CanisterId,
@@ -474,7 +468,6 @@ impl Data {
         user_index_canister_id: CanisterId,
         local_user_index_canister_id: CanisterId,
         group_index_canister_id: CanisterId,
-        local_group_index_canister_id: CanisterId,
         notifications_canister_id: CanisterId,
         proposals_bot_user_id: UserId,
         escrow_canister_id: CanisterId,
@@ -521,7 +514,6 @@ impl Data {
             user_index_canister_id,
             local_user_index_canister_id,
             group_index_canister_id,
-            local_group_index_canister_id,
             notifications_canister_id,
             proposals_bot_user_id,
             escrow_canister_id,
@@ -547,7 +539,7 @@ impl Data {
             total_payment_receipts: PaymentReceipts::default(),
             video_call_operators,
             ic_root_key,
-            event_store_client: EventStoreClientBuilder::new(local_group_index_canister_id, CdkRuntime::default())
+            event_store_client: EventStoreClientBuilder::new(local_user_index_canister_id, CdkRuntime::default())
                 .with_flush_delay(Duration::from_millis(5 * MINUTE_IN_MS))
                 .build(),
             achievements: Achievements::default(),
@@ -563,7 +555,7 @@ impl Data {
             notifications_queue: BatchedTimerJobQueue::new(
                 NotificationPusherState {
                     notifications_canister: notifications_canister_id,
-                    authorizer: local_group_index_canister_id,
+                    authorizer: local_user_index_canister_id,
                 },
                 false,
             ),
@@ -1102,7 +1094,6 @@ pub struct CanisterIds {
     pub user_index: CanisterId,
     pub group_index: CanisterId,
     pub local_user_index: CanisterId,
-    pub local_group_index: CanisterId,
     pub notifications: CanisterId,
     pub proposals_bot: CanisterId,
     pub escrow: CanisterId,