+++ backend/canisters/group/impl/src/updates/c2c_local_group_index.rs
@@ -1,5 +1,5 @@
 use crate::activity_notifications::handle_activity_notification;
-use crate::guards::{caller_is_local_group_index, caller_is_local_user_index};
+use crate::guards::caller_is_local_user_index;
 use crate::{RuntimeState, mutate_state, run_regular_jobs};
 use canister_api_macros::update;
 use canister_time::now_millis;
@@ -16,22 +16,14 @@ use types::{GroupNameChanged, TimestampMillis, Timestamped};
 fn c2c_local_index(args: Args) -> Response {
     run_regular_jobs();
 
-    mutate_state(|state| c2c_local_group_index_impl(args, state))
+    mutate_state(|state| c2c_local_index_impl(args, state))
 }
 
-#[update(guard = "caller_is_local_group_index", msgpack = true)]
-#[trace]
-fn c2c_local_group_index(args: Args) -> Response {
-    run_regular_jobs();
-
-    mutate_state(|state| c2c_local_group_index_impl(args, state))
-}
-
-fn c2c_local_group_index_impl(args: Args, state: &mut RuntimeState) -> Response {
+fn c2c_local_index_impl(args: Args, state: &mut RuntimeState) -> Response {
     let now = LazyCell::new(now_millis);
     for event in args.events {
         if state.data.idempotency_checker.check(
-            state.data.local_group_index_canister_id,
+            state.data.local_user_index_canister_id,
             event.created_at,
             event.idempotency_id,
         ) {