+++ backend/canisters/local_user_index/api/src/lib.rs
@@ -50,6 +50,27 @@ pub enum UserIndexEvent {
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
+pub enum GroupIndexEvent {
+    GroupNameChanged(NameChanged),
+    CommunityNameChanged(NameChanged),
+    GroupVerifiedChanged(VerifiedChanged),
+    CommunityVerifiedChanged(VerifiedChanged),
+    NotifyOfUserDeleted(CanisterId, UserId),
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub struct NameChanged {
+    pub canister_id: CanisterId,
+    pub name: String,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub struct VerifiedChanged {
+    pub canister_id: CanisterId,
+    pub verified: bool,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
 pub enum LocalGroupIndexEvent {
     MigrateGroup(ChatId, LocalGroup),
     MigrateCommunity(CommunityId, LocalCommunity),