+++ backend/canisters/local_user_index/impl/src/guards.rs
@@ -8,6 +8,14 @@ pub fn caller_is_user_index_canister() -> Result<(), String> {
     }
 }
 
+pub fn caller_is_user_index_or_group_index() -> Result<(), String> {
+    if read_state(|state| state.is_caller_user_index_canister() || state.is_caller_group_index_canister()) {
+        Ok(())
+    } else {
+        Err("Caller is not the UserIndex or GroupIndex".to_string())
+    }
+}
+
 pub fn caller_is_group_index_canister() -> Result<(), String> {
     if read_state(|state| state.is_caller_group_index_canister()) {
         Ok(())